requirements.txt
pytest==6.2.5
argparse==1.4.0
pytest-mock==0.6.0
pyyaml==6.0
pycryptodome==3.14.1
pytest-aiohttp==1.0.4
psycopg2-binary==2.8.6